KSUIE Economy Push
My first release: These scripts create a Butcher Shop, a Gold Shop, and a Supply Shop where players can buy or sell items. It also includes an economy to track items bought and sold and raise or lower prices based on volumes of transactions.  The economy is kept balanced by the server consuming 50 of all items every 10 minutes to ensure prices don't dip too low and to boost the economy when the server is inactive to encourage players to return and check prices.